pyproject.toml
[tool.black]
line-length = 100
target-version = ['py37']

[tool.isort]
known_third_party = [
  "black",
  "click",
  "click_option_group",
  "hydra",
  "matplotlib",
  "mypy",
  "numpy",
  "omegaconf",
  "pandas",
  "pylint",
  "pytest",
  "scipy",
  "setuptools",
  "sklearn",
  "torch",
  "tqdm",
  "wandb",
]
known_future_library = []
extra_standard_library = ["dataclasses", "__future__"]
line_length = 88
profile = "black"
force_sort_within_sections = "True"
classes = ["MISSING"]

[tool.poetry]
name = "landscapes"
version = "0.0.1"
description = ""
authors = ["Myles Bartlett"]
license = "Apache License 2.0"
packages = [
  { include = "landscapes" },
  { include = "hydra_plugins" },
  { include = "hydra_plugins/submitit_launcher" },
]
include=["ranzen/py.typed"]
classifiers=[
  "Programming Language :: Python :: 3.9",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Typing :: Typed",
]
keywords=["typing", "python"]
repository=""
readme="README.md"

[tool.poetry.dependencies]
python = ">=3.9.1,<3.11"
torch = "^1.10.2"
ranzen = {version = "^1.2.1", extras = ["all"]}
conduit = { git = "https://github.com/wearepal/conduit.git", branch = "main" }
wandb = "^0.12.10"
attrs = "^21.4.0"
torchmetrics = "^0.7.2"
regex = "^2022.1.18"
ftfy = "^6.1.1"
clip = { git = "https://github.com/openai/CLIP.git", branch = "main" }
classy-vision = "^0.6.0"
setuptools = "59.5.0" # for pytorch-lightning
fairscale = "^0.4.5"
hydra-submitit-launcher = "^1.1.6"
typing-extensions = ">=4.0"
scann = { version = "^1.2.5", platform = "linux" }
deepspeed = { version = "^0.5.10", platform = "linux" }
triton = { version = "1.0.0", platform = "linux" }
torchvision = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^22.1"
mypy = "^0.800"
pytest = "^6.2.2"
isort = "^5.7.0"
darglint = "^1.8.0"

[tool.pyright]
exclude = ["**/node_modules", "**/__pycache__"]
typeCheckingMode = "basic"
pythonVersion = "3.9"
reportUnusedImport = "error"
reportDuplicateImport  = "error"
reportIncompatibleVariableOverride = "error"
reportOverlappingOverload = "error"
reportUntypedNamedTuple = "error"
reportConstantRedefinition = "warning"
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
strictListInference = true
strictSetInference = true
strictParameterNoneValue = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
